Refactor unit test and create BaseUnitTestCase
