+++ tests/Unit/Detector/MySQLGoneAwayDetectorTest.php
@@ -3,15 +3,10 @@
 namespace Facile\DoctrineMySQLComeBack\Tests\Unit\Detector;
 
 use Facile\DoctrineMySQLComeBack\Doctrine\DBAL\Detector\MySQLGoneAwayDetector;
-use Facile\DoctrineMySQLComeBack\Tests\DeprecationTrait;
-use PHPUnit\Framework\TestCase;
-use Prophecy\PhpUnit\ProphecyTrait;
+use Facile\DoctrineMySQLComeBack\Tests\Unit\BaseUnitTestCase;
 
-class MySQLGoneAwayDetectorTest extends TestCase
+class MySQLGoneAwayDetectorTest extends BaseUnitTestCase
 {
-    use ProphecyTrait;
-    use DeprecationTrait;
-
     private const RETRYABLE_ERROR = 'Lost connection to MySQL server during query is an error not retryable on UPDATE queries';
 
     private const RETRYABLE_ERROR_OUTSIDE_UPDATE = 'Lost connection to MySQL server during query is an error not retryable on UPDATE queries';
